+++ src/main/java/com/ssafy/soljigi/api/service/AccountService.java
@@ -2,7 +2,7 @@ package com.ssafy.soljigi.api.service;
 
 import org.springframework.stereotype.Service;
 
-import com.ssafy.soljigi.api.dto.AccountTransactionDto;
+import com.ssafy.soljigi.api.dto.response.AccountTransactionResponse;
 import com.ssafy.soljigi.api.entity.Account;
 import com.ssafy.soljigi.api.repository.AccountRepository;
 
@@ -18,9 +18,9 @@ public class AccountService {
 		accountRepository.save(account);
 	}
 
-	public AccountTransactionDto get(String accountNumber) {
+	public AccountTransactionResponse get(String accountNumber) {
 		Account account = accountRepository.findByAccountNumber(accountNumber)
 			.orElseThrow(IllegalArgumentException::new);
-		return AccountTransactionDto.of(account);
+		return AccountTransactionResponse.of(account);
 	}
 }
